Add files via upload
The code defines two regular expressions and uses the Python re module to perform lexical analysis by extracting the valid tokens. The ast module is used to construct a syntax tree for the given regular expressions, and then the tree is printed using ast.dump() function. Finally, the syntax trees are executed using the exec() function. The lexical analyzer and syntax tree functions are called, and the output of the extracted tokens and syntax tree nodes are printed.